Remove check_overcr
This commit removes the plugin check_overcr and all
related files and parts of files.

OverCR seems to be dead and I have never seen a mention of
an active usage anywhere.
The website still does exist, but the activity seems to be
limited.
Carrying check_overcr forward seems to be a burden without a
purpose and more of an opportunity to remove some code.