append the query string from parsed uri
Check the UriUriA object, and if query string exists append it to the
new_url. Only appends the query part, fragments are still not appended

Function redir parses the new location header value using the
uriParseUriA function already, which populates the query field. This
field was already being printed, but it was not being appended to the
new_url during its construction.

Redirection chain of check_curl --onredirect=follow now mimics the chain
of check_http --onredirect=follow. Tested on the url:
mail.google.com/chat